+++ delta-coverage-gradle/src/test/kotlin/io/github/surpsg/deltacoverage/gradle/sources/lookup/JacocoPluginSourcesLookupTest.kt
@@ -37,7 +37,7 @@ internal class JacocoPluginSourcesLookupTest {
             SourcesAutoLookup.Context(
                 project = project,
                 viewName = JavaPlugin.TEST_TASK_NAME,
-                deltaCoverageConfiguration = DeltaCoverageConfiguration(project, project.objects),
+                deltaCoverageConfiguration = DeltaCoverageConfiguration(project.objects),
                 objectFactory = project.objects,
             )
         )
@@ -60,7 +60,7 @@ internal class JacocoPluginSourcesLookupTest {
             SourcesAutoLookup.Context(
                 project = project,
                 viewName = JavaPlugin.TEST_TASK_NAME,
-                deltaCoverageConfiguration = DeltaCoverageConfiguration(project, project.objects),
+                deltaCoverageConfiguration = DeltaCoverageConfiguration(project.objects),
                 objectFactory = project.objects,
             )
         )